caption_requirements.txt
transformers
Pillow
numpy
scikit-learn
tqdm
tyro
lmdeploy
